requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
scipy>=1.11.0
aiohttp>=3.9.0
py-clob-client>=0.24.0
py-builder-relayer-client>=0.3.0
py-builder-signing-sdk>=0.3.0
pytest>=9.0.0
